requirements.txt
fastapi[standard]==0.112.2
httpx==0.27.2
prometheus_client==0.20.0
python-dotenv==1.0.1